network: Moved network generation to a header-only file
